🏗️ Reorganize CLIEngine structure: 'widgets' header include `Choices` and `Dice`
